+++ qcmanybody/models/manybody_output_pydv1.py
@@ -312,6 +312,37 @@ else:
     )
 
 
+def _qcvars_translator(cls, reverse: bool = False) -> Dict[str, str]:
+    """Form translation map between many-body results QCSchema and Psi4/QCDB terminologies.
+
+    Parameters
+    ----------
+    reverse
+        Keys are QCVariable names (`reverse=True`) rather than QCSchema names (default; `reverse=False`).
+
+    Returns
+    -------
+    dict
+        Map from ManyBodyResultProperties field names to QCVariable names, or reverse.
+
+    """
+    qcvars_to_mbprop = {}
+    # v2: for skprop in ManyBodyResultProperties.model_fields.keys():
+    for skprop in cls.__fields__.keys():
+        qcvar = skprop.replace("_body", "-body").replace("_corr", "-corr").replace("_", " ").upper()
+        qcvars_to_mbprop[qcvar] = skprop
+    for ret in ["energy", "gradient", "hessian"]:
+        qcvars_to_mbprop[f"CURRENT {ret.upper()}"] = f"return_{ret}"
+
+    if reverse:
+        return qcvars_to_mbprop
+    else:
+        return {v: k for k, v in qcvars_to_mbprop.items()}
+
+
+ManyBodyResultProperties.to_qcvariables = classmethod(_qcvars_translator)
+
+
 # ====  Results  ================================================================
 
 
@@ -353,7 +384,7 @@ class ManyBodyResult(SuccessfulResultBase):
         description="The primary logging output of the program, whether natively standard output or a file. Presence vs. absence (or null-ness?) configurable by protocol.",
     )
     stderr: Optional[str] = Field(None, description="The standard error of the program execution.")
-    success: Literal[True] = Field(True, description="Always `True` for a successful result")
+    # v2: success: Literal[True] = Field(True, description="Always `True` for a successful result")
 
     @validator("component_results")
     def _component_results(cls, value, values):